+++ blockchain/chain/src/store/chain_store.rs
@@ -18,8 +18,8 @@ use crossbeam::atomic::AtomicCell;
 use crypto::DomainSeparationTag;
 use encoding::{blake2b_256, de::DeserializeOwned, from_slice, Cbor};
 use forest_car::CarHeader;
-use forest_ipld::Ipld;
 use futures::{AsyncRead, AsyncWrite};
+use forest_ipld::{recurse_links, Ipld};
 use interpreter::BlockMessages;
 use ipld_amt::Amt;
 use ipld_blockstore::BlockStore;
@@ -750,63 +750,6 @@ where
         }
         Ok(())
     }
-}
-
-// Traverses all Cid links, loading all unique values and using the callback function
-// to interact with the data.
-fn traverse_ipld_links<F>(
-    walked: &mut HashSet<Cid>,
-    load_block: &mut F,
-    ipld: &Ipld,
-) -> Result<(), Box<dyn StdError>>
-where
-    F: FnMut(Cid) -> Result<Vec<u8>, Box<dyn StdError>>,
-{
-    match ipld {
-        Ipld::Map(m) => {
-            for (_, v) in m.iter() {
-                traverse_ipld_links(walked, load_block, v)?;
-            }
-        }
-        Ipld::List(list) => {
-            for v in list.iter() {
-                traverse_ipld_links(walked, load_block, v)?;
-            }
-        }
-        Ipld::Link(cid) => {
-            if cid.codec() == cid::DAG_CBOR {
-                if !walked.insert(*cid) {
-                    return Ok(());
-                }
-                let bytes = load_block(*cid)?;
-                let ipld = Ipld::unmarshal_cbor(&bytes)?;
-                traverse_ipld_links(walked, load_block, &ipld)?;
-            }
-        }
-        _ => (),
-    }
-    Ok(())
-}
-
-// Load cids and call [traverse_ipld_links] to resolve recursively.
-fn recurse_links<F>(walked: &mut HashSet<Cid>, root: Cid, load_block: &mut F) -> Result<(), Error>
-where
-    F: FnMut(Cid) -> Result<Vec<u8>, Box<dyn StdError>>,
-{
-    if !walked.insert(root) {
-        // Cid has already been traversed
-        return Ok(());
-    }
-    if root.codec() != cid::DAG_CBOR {
-        return Ok(());
-    }
-
-    let bytes = load_block(root)?;
-    let ipld = Ipld::unmarshal_cbor(&bytes)?;
-
-    traverse_ipld_links(walked, load_block, &ipld)?;
-
-    Ok(())
 }
 
 pub(crate) type TipsetCache = RwLock<LruCache<TipsetKeys, Arc<Tipset>>>;
